requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.0.8
django-environ==0.4.5
djangorestframework==3.11.0
flake8==3.8.3
idna==2.10
isort==5.0.3
mccabe==0.6.1
pycodestyle==2.6.0
pyflakes==2.2.0
python-dotenv==0.14.0
pytz==2020.1
sqlparse==0.3.1
urllib3==1.25.9
